Back-End/requirements.txt
Flask
python-dotenv
openai-agents
asyncio
uvicorn 
asgiref
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2
sqlalchemy==2.0.25
Flask-SQLAlchemy==3.1.1
tiktoken


